% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment_functions.R
\name{check_assessment}
\alias{check_assessment}
\title{Check whether the assessments have converged}
\usage{
check_assessment(assessment_ob, save_result = FALSE)
}
\arguments{
\item{assessment_ob}{A HARSAT assessment object resulting from a call to
ctsm_assessment}

\item{save_result}{Saves the identifiers of the timeseries that have not
converged; defaults to FALSE. When an assessment has been done in stages, the output also identifies those
timeseries that have not yet been assessed}
}
\value{
A list of two character vectors:
\itemize{
\item not_converged identifies timeseries that have not converged
\item not_assessed identifies timeseries that have not been assessed
}
}
\description{
Checks whether the assessments in a HARSAT assessment object
have converged. Currently only does detailed checks for models with normal
or lognormal errors (all chemical timeseries and some biological effects).
Here it checks whether:
\itemize{
\item fixed effect estimates are away from their bounds
\item random effect estimates are lower than their upper bounds; they can of
course be equal to zero
\item standard errors are present for model predictions and fixed effects
estimates
\item standard errors of the fixed effects estimates are realistic (very small standard errors indicate problems with
the numerical differencing used to compute the standard errors)
}
}
