% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_functions.R
\name{convert_reftable}
\alias{convert_reftable}
\title{Add OSPAR_subregion to simplified sediment threshold reference table}
\usage{
convert_reftable(input_file, output_file, export = TRUE)
}
\arguments{
\item{input_file}{the input reference file}

\item{output_file}{the expanded reference file}

\item{export}{a boolean flag, if \code{FALSE}, the data is returned rather than being written to \code{output_file}}
}
\value{
if \code{export} is \code{FALSE} (the default), returns the expanded data
}
\description{
This is a utility function to expand a simplified OSPAR sediment threshold
table (which is much easier for the user to edit) into the form required by
\code{harsat}
}
