% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_functions.R
\name{convert_units}
\alias{convert_units}
\title{Converts units, e.g., from mg/kg to ug/kg}
\usage{
convert_units(conc, from, to)
}
\arguments{
\item{conc}{the value to convert}

\item{from}{the current units}

\item{to}{the required units}
}
\description{
Can accept non-standard units (e.g. for biological effects) provided that
\code{from} and \code{to} for these rows are identical (in which case no attempt is made
to convert.
}
