% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{create_timeseries}
\alias{create_timeseries}
\title{Create a time series}
\usage{
create_timeseries(
  ctsm.obj,
  determinands = ctsm_get_determinands(ctsm.obj$info),
  determinands.control = NULL,
  oddity_path = "oddities",
  return_early = FALSE,
  print_code_warnings = FALSE,
  get_basis = get_basis_default,
  normalise = FALSE,
  normalise.control = list()
)
}
\arguments{
\item{ctsm.obj}{the harsat object, as returned from \code{tidy_data}}

\item{determinands}{the determinands to use, by default derived by
calling \code{ctsm_get_determinands}, which takes values from
the determinand reference table}

\item{determinands.control}{determinands control values, when needed}

\item{oddity_path}{a directory to write the oddities to}

\item{return_early}{a boolean (default \code{FALSE}), if \code{TRUE}, returns early}

\item{print_code_warnings}{a boolean (default \code{FALSE})}

\item{get_basis}{a basis function, by default \code{\link{get_basis_default}}}

\item{normalise}{boolean or function, if TRUE, uses a default normalization}

\item{normalise.control}{a list of control data for the normalization function}
}
\value{
a completed timeseries object, which can be used for assessments
}
\description{
Cleans the data and turns it into time series structures ready for assessment
}
