% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportional_odds_functions.R
\name{cstm.VDS.environment}
\alias{cstm.VDS.environment}
\title{Detects the environment from the call}
\usage{
cstm.VDS.environment()
}
\description{
This is a utility function that detects the package environment.
If you have imported \code{harsat} as a package, it returns the package
environment. Otherwise, it returns the global environment. You can safely
export functions from the result of this, for example, when you are
setting up a cluster of child processes for parallel computation.
}
