% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_functions.R
\name{ctsm.projection}
\alias{ctsm.projection}
\title{Calculates a polar projection}
\usage{
ctsm.projection(latitude, longitude)
}
\arguments{
\item{latitude}{the latitude}

\item{longitude}{the longitude}
}
\value{
a list with projected \code{longitude} (easting) and \code{latitude} (northing) values
}
\description{
Calculates easting and northing in Lambert Azimuthal Equal
Area North Polar aspect projection
}
