% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportional_odds_functions.R
\name{ctsm.VDS.loglik.calc}
\alias{ctsm.VDS.loglik.calc}
\title{ctsm.VDS.loglik.calc}
\usage{
ctsm.VDS.loglik.calc(
  theta,
  data,
  index.theta,
  minus.twice = FALSE,
  cumulate = FALSE
)
}
\arguments{
\item{theta}{The maximum imposex stage}

\item{data}{Individual imposex data.}

\item{index.theta}{Allows for stage names that do not run from 0 to \code{theta}
(optional)}

\item{minus.twice}{A logical specifying whehter to calculated the liklihood
(FALSE) or the deviance (TRUE); default FALSE}

\item{cumulate}{A logical specifying whether to use cumulative probabilities;
default FALSE}
}
\description{
ctsm.VDS.loglik.calc
}
