% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportional_odds_functions.R
\name{ctsm.VDS.p.calc}
\alias{ctsm.VDS.p.calc}
\title{ctsm.VDS.p.calc}
\usage{
ctsm.VDS.p.calc(theta, cumulate = FALSE)
}
\arguments{
\item{cumulate}{a boolean, whether to use cumulative probabilities}
}
\description{
ctsm.VDS.p.calc
}
