% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportional_odds_functions.R
\name{ctsm.VDS.varlist}
\alias{ctsm.VDS.varlist}
\title{ctsm.VDS.varlist}
\usage{
ctsm.VDS.varlist
}
\description{
A list of names for functions and values necessary to export to
cluster prcesses for parallel computation
}
