% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment_functions.R
\name{ctsm_check_convergence}
\alias{ctsm_check_convergence}
\title{Check for convergence}
\usage{
ctsm_check_convergence(assessment_ob, coeff_se_tol = 0.001)
}
\description{
Check for convergence
}
