% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_functions.R
\name{ctsm_convert_basis}
\alias{ctsm_convert_basis}
\title{Basis and matrix information and basis conversion}
\usage{
ctsm_convert_basis(
  conc,
  from,
  to,
  drywt = NA_real_,
  lipidwt = NA_real_,
  drywt_censoring = ifelse(is.na(drywt), NA_character_, ""),
  lipidwt_censoring = ifelse(is.na(lipidwt), NA_character_, ""),
  exclude = FALSE,
  print_warning = TRUE
)
}
\arguments{
\item{conc}{the concentration}

\item{from}{the current basis}

\item{to}{the target basis}

\item{drywt}{assumed to be a percentage taking values in [0, 100]}

\item{lipidwt}{assumed to be a percentage taking values in [0, 100];
assumed to be on a wet weight basis}

\item{drywt_censoring}{required when \code{drywt} is present,
character: must be '', '<', 'D', 'Q' or NA}

\item{lipidwt_censoring}{required when \code{lipidwt} is present,
character: must be '', '<', 'D', 'Q' or NA}

\item{exclude}{a logical identifying records that do not need to be converted,
useful when the data contain biological effects measurements}

\item{print_warning}{a boolean, gives the number of records lost during conversion}
}
\description{
Converts concentrations between wet, dry and lipid bases
}
