% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_functions.R
\name{ctsm_get_determinands}
\alias{ctsm_get_determinands}
\title{Get the determinands}
\usage{
ctsm_get_determinands(info)
}
\description{
Gets the determinands to be assessed from the determinand reference table
}
