% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_functions.R
\name{ctsm_get_info}
\alias{ctsm_get_info}
\title{Gets data from the standard reference tables}
\usage{
ctsm_get_info(
  ref_table,
  input,
  output,
  compartment = NULL,
  na_action = c("fail", "input_ok", "output_ok", "ok"),
  info_type = NULL,
  sep = "."
)
}
\description{
Gets data from the standard reference tables
}
