% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_functions.R
\name{ctsm_plot_assessment}
\alias{ctsm_plot_assessment}
\title{Plot the assessment}
\usage{
ctsm_plot_assessment(
  ctsm_asmt_obj,
  subset = NULL,
  output_dir = ".",
  file_type = c("data", "index"),
  file_format = c("png", "pdf")
)
}
\description{
Generates a series of assessment plots with the raw data, or the annual
indices, or both
}
