% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{ctsm_read_data}
\alias{ctsm_read_data}
\title{Read harsat data}
\usage{
ctsm_read_data(
  compartment = c("biota", "sediment", "water"),
  purpose = c("OSPAR", "HELCOM", "AMAP", "other"),
  contaminants,
  stations,
  QA,
  data_dir = ".",
  info_files = list(),
  info_dir = ".",
  extraction = NULL,
  max_year = NULL,
  oddity_dir = "oddities",
  control = list(),
  data_format = c("ICES_old", "ICES_new", "external")
)
}
\arguments{
\item{compartment}{a string, one of "biota", "sediment", and "water"}

\item{purpose}{a string, one of "OSPAR", "HELCOM", "AMAP" and "other"}

\item{contaminants}{a file reference for contaminant data}

\item{stations}{a file reference for station data}

\item{QA}{a file reference for QA data}

\item{data_dir}{The directory where the data files can be found (sometimes
supplied using 'file.path'). Defaults to "."; i.e. the working directory.}

\item{info_files}{A list of files specifying reference tables which override
the defaults. See examples.}

\item{info_dir}{The directory where the reference tables can be found
(sometimes supplied using 'file.path'). Defaults to "."; i.e. the working
directory}

\item{extraction}{A date saying when the extraction was made. Optional. This
should be provided according to ISO 8601; for example, 29 February 2024
should be supplied as "2024-02-29".}

\item{oddity_dir}{The directory where the 'oddities' will be written
(sometimes supplied using 'file.path'). This directory (and subdirectories)
will be created if it does not already exist.}

\item{data_format}{a string, one of "ICES_old", "ICES_new", and "external"}
}
\description{
Read harsat data
}
