% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_functions.R
\name{ctsm_summary_table}
\alias{ctsm_summary_table}
\title{Generates a summary table}
\usage{
ctsm_summary_table(
  assessment_obj,
  determinandGroups = NULL,
  classColour = NULL,
  output_dir = NULL,
  output_file = NULL,
  export = TRUE,
  collapse_AC = NULL
)
}
\arguments{
\item{output_file}{the output file -- it must be a CSV file}
}
\description{
Creates a summary file and adds symbology
}
