% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{ctsm_TBT_convert}
\alias{ctsm_TBT_convert}
\title{Convert tin concentrations}
\usage{
ctsm_TBT_convert(
  data,
  subset,
  action,
  from = c("tin", "cation"),
  convert_var = c("value", "limit_detection", "limit_quantification", "uncertainty")
)
}
\description{
Convert tin concentrations to cation concentrations.
Also \code{change_unit} moves units from tin units to conventional units
}
