% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{ctsm_tidy_data}
\alias{ctsm_tidy_data}
\title{Tidy the data}
\usage{
ctsm_tidy_data(ctsm_obj)
}
\arguments{
\item{ctsm_obj}{}
}
\description{
Reduces the size of the extraction by removing redundant variables.
Any ad-hoc changes will usually be made between \code{read_data} and \code{simplify_data}.
The output is in the correct format for \code{create_timeSeries}.
}
