% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty_functions.R
\name{ctsm_uncrt_plot_data}
\alias{ctsm_uncrt_plot_data}
\title{Generates a relative uncertainty plot}
\usage{
ctsm_uncrt_plot_data(uncrt_obj, det_id)
}
\arguments{
\item{uncrt_obj}{an uncertainty object}

\item{det_id}{a list of determinand ids}
}
\description{
Generates a relative uncertainty plot
}
