% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_functions.R
\docType{data}
\name{get_AC}
\alias{get_AC}
\title{Access function map}
\format{
An object of class \code{list} of length 3.
}
\usage{
get_AC
}
\description{
Access function map
}
\keyword{datasets}
