% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_functions.R
\name{get_AC_biota_OSPAR}
\alias{get_AC_biota_OSPAR}
\title{Gets OSPAR threshold values for biota}
\usage{
get_AC_biota_OSPAR(data, AC, rt, export_all = FALSE)
}
\arguments{
\item{data}{the data}

\item{AC}{a list of assessment criteria}

\item{rt}{the reference tables}

\item{export_all}{logical, default \code{FALSE}, if \code{TRUE} returns all data,
otherwise, just the assessment criteria}
}
\value{
by default, the assessment criteria, if \code{export_all} is set \code{TRUE}
returns all the data instead
}
\description{
Extends default extractor function get_AC$biota for the few cases which can
not be handled in a straightforward manner. This mostly relates to thresholds
which are only applied when the typical species / lipid contend is 'high'
(currently defined as >= 3\%)
}
