% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_functions.R
\name{get_basis_biota_OSPAR}
\alias{get_basis_biota_OSPAR}
\title{Gets the OSPAR biota target basis}
\usage{
get_basis_biota_OSPAR(data, info)
}
\arguments{
\item{data}{the data}

\item{info}{the information object}
}
\description{
Gets the OSPAR biota target basis
}
