% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{get_control_dets}
\alias{get_control_dets}
\title{extract determinands from control structure}
\usage{
get_control_dets(control, .names = TRUE)
}
\arguments{
\item{control}{list of control structures passed into \code{create_timeseries} by
argument \code{determinand.control}}

\item{.names}{logical determining whether the names of the list are included
(\code{TRUE} default) or not (\code{FALSE})}
}
\value{
charcater string of determinands
}
\description{
gets the names of all determinands involved in argument \code{determinands.control}
}
