% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_functions.R
\name{get_determinands}
\alias{get_determinands}
\title{Get the determinands to be assessed}
\usage{
get_determinands(info)
}
\arguments{
\item{info}{A list with at least the following two components:
\itemize{
\item \code{compartment} One of \code{"biota"}, \code{"sediment"} or \code{"water"}
\item \code{determinand} A data frame holding the determinand reference table
}}
}
\value{
A character string containing the determinands to be assessed. The
function will fail with an error message if there are no such determinands.
}
\description{
Gets the determinands to be assessed from the determinand reference table.
}
\details{
The determinands are taken from the column \code{biota_assess},
\code{sediment_assess} or \code{water_assess} in the determinand reference table
(where the compartment is given by \code{info$compartment}). \code{TRUE} values are
determinands that are to be assessed.
}
