% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_functions.R
\name{get_station_code}
\alias{get_station_code}
\title{Get station code from station name}
\usage{
get_station_code(station_name, country, stations)
}
\description{
Gets the station code corresponding to the station name and country from the
station dictionary. Only works for one country at a time
}
