% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harsat-package.R
\docType{package}
\name{harsat-package}
\alias{harsat}
\alias{harsat-package}
\title{harsat: Harmonized Regional Seas Assessment Tool}
\description{
Tools for the assessment of data concerning contaminants (hazardous substances) and their effects in the marine environment. The code includes tools for pre-processing data, statistical trend analysis and comparison with threshold values, and post-processing for archiving and reporting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/osparcomm/HARSAT}
  \item \url{https://osparcomm.github.io/HARSAT/}
  \item Report bugs at \url{https://github.com/osparcomm/HARSAT/issues}
}

}
\author{
\strong{Maintainer}: OSPAR Commission (OSPAR) \email{data@ospar.org} [copyright holder, funder]

Authors:
\itemize{
  \item OSPAR Commission (OSPAR) \email{data@ospar.org} [copyright holder, funder]
  \item Arctic Monitoring and Assessment Programme (AMAP) \email{amap@amap.no} [copyright holder, funder]
  \item Helsinki Commission (HELCOM) \email{secretariat@helcom.fi} [copyright holder, funder]
  \item International Council for the Exploration of the Sea (ICES) \email{info@ices.dk} [copyright holder]
  \item AmbieSense Ltd \email{info@ambiesense.com} [copyright holder]
}

Other contributors:
\itemize{
  \item HARSAT User Group \email{harsat@ospar.org} [contributor]
}

}
\keyword{internal}
