% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_functions.R
\name{label_normaliser}
\alias{label_normaliser}
\title{Text representation of normaliser information for plots and reports}
\usage{
label_normaliser(normaliser, value, unit, html = FALSE)
}
\arguments{
\item{normaliser}{Character vector giving the names of the normaliser (ICES
code rather than common name)}

\item{value}{Numeric vector giving the value of the normaliser}

\item{unit}{Character vector giving the unit of the normaliser}

\item{html}{Logical with TRUE returning an html representation for use in
markdown and FALSE (default) returning an text expression for use in lattice
(grid) graphics}
}
\value{
A character string that can be used in markdown (\code{html = TRUE}) or used in
grid graphics ('html = FALSE\verb{). In the latter, need to turn the string into an expression by }parse(text = result)`.
}
\description{
Utilty function for providing inforamtion about normalistation to plots in
e.g. \code{plot_assessment} (where it is used with \code{label_unit}) or
html reports in e.g. \code{report_assessment}.
}
\examples{
label_normaliser("AL", 2.5, "\%")
label_normaliser("AL", 2.5, "\%", html = TRUE)

}
