% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_functions.R
\name{label_unit}
\alias{label_unit}
\title{Text representation of units for plots and reports}
\usage{
label_unit(
  unit,
  basis = NA_character_,
  normaliser = NA_character_,
  normaliser_value = NA_real_,
  normaliser_unit = NA_character_,
  html = FALSE
)
}
\arguments{
\item{unit}{Character vector giving the units}

\item{basis}{Character vector giving the bases. Must be one of "D", "L",
"W" or NA_character_}

\item{normaliser}{Character vector giving the names of the normaliser (if
there are any)}

\item{normaliser_value}{Numeric vector giving the values of the normaliser
(if there are any)}

\item{normaliser_unit}{Character vector giving the units of the normaliser
(if there are any)}

\item{html}{Logical with TRUE returning an html representation for use in
markdown and FALSE (default) returning an text expression for use in lattice
(grid) graphics}
}
\value{
A character vector that can be used in markdown (\code{html = TRUE}) or in
grid graphics ('html = FALSE\verb{). In the latter, need to turn the strings into an expression by }parse(text = result)`.
}
\description{
Utilty function for adding units to plots in e.g. \code{plot_assessment} or
html reports in e.g. \code{report_assessment}. Standard (recognised) units are
prettified, with non-standard units returned unchanged.
}
