% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{locate_information_file}
\alias{locate_information_file}
\title{Find the path for an information file}
\usage{
locate_information_file(path, name)
}
\arguments{
\item{path}{A string giving the directory to search. The information directory
for the package is automatically searched if we haven't found the
file in this directory}

\item{name}{A string: the name of the file, e.g., \code{thresholds_biota.csv}}
}
\value{
A string, the path to the file, or \code{""} if the file cannot be
found.
}
\description{
Locates a requested information file, searching the information file path,
and then looking in the package itself.
}
