% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_functions.R
\name{make_key}
\alias{make_key}
\title{Make key for annotating assessment plots}
\usage{
make_key(series, info, type)
}
\arguments{
\item{series}{List describing the timeseries in the plot. If there are
multiple timeseries, then some entries of the list require multiple entries,
one for each timeseries. See details.}

\item{info}{The standard \code{info} element of the harsat object.}

\item{type}{Character scalar (default \code{"data"}) describing the type of plot:
other options are \code{"index"}, \code{"auxiliary"}, \code{"data_splom"}, \code{"index_mp"},
\code{"ratio_mp"}.}
}
\value{
a list with four text (or expression) elements:
\itemize{
\item \code{media}: compartment , species and matrix (biota), matrix (sediment),
filtration (water), subseries
\item \code{station}: station_name and station_longname
\item \code{units}: an expression giving the units including any normalisation
\item \code{extraction}: the extraction date
}

Both \code{media} and \code{units} can be complicated if there are multiple
timeseries, particularly if e.g. there are different normalisers (sediment)
or matrices (biological effects) involved
`
}
\description{
Provides the meta-data that accompanies each plot in \code{plot_assessment}:
compartment, station, units and data extraction. It also provides extended
information for plots involving multiple timeseries e.g. in \code{plot_multidata},
\code{plot_multiassessment} and \code{plot_ratio}.
}
\details{
\code{series} is a list with giving information about a single timeseries
(when called by \code{plot_assessment}) or multiple related timeseries (when
called by \code{plot_multidata}, \code{plot_multiassessment}, and \code{plot_ratio}). It
must contain the following (compartment-specific) elements:
\itemize{
\item \code{determinand} (all): a character vector giving the
determinand for each timeseries; if there are multiple timeseries, then the
vector must have names which are the corresponding series identifiers
\item \code{matrix} (biota, sediment): a character vector giving the matrix for
each timeseries; if there are multiple timeseries, then the vector must have
names which are the corresponding series identifiers;
the matrices might vary across timeseries, but this is currently likely only
for biological effects
\item \code{filtration} (water): a character vector giving the filtration method for
each timeseries; if there are multiple timeseries, then the vector must have
names which are the corresponding series identifiers;
the filtration methods might vary across timeseries, but this is currently
unlikely in practice
\item \code{basis} (all): a character vector giving the basis for
each timeseries; if there are multiple timeseries, then the vector must have
names which are the corresponding series identifiers;
the bases might vary across timeseries, but this is currently likely only for
biological effects
\item \code{species} (biota): a character giving the species (common to all
timeseries)
\item \code{subseries} (all): a character giving the subseries (common to all
timeseries); can be \code{NA_character_}
}

In addition, if \code{info$normalise} is \code{TRUE}, then it must also contain:
\itemize{
\item \code{normaliser}, \code{normaliser_value} and \code{normaliser_unit}: character, numeric,
and character vectors giving the normaliser, it's value and the unit on which
it is expressed for each timeseries; missing values are allowed if there is
no normalisation for the timeseries involved
}
}
