% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{merge_auxiliary}
\alias{merge_auxiliary}
\title{Merge auxiliary variables with data}
\usage{
merge_auxiliary(data, info, determinands)
}
\arguments{
\item{data}{a data frame containing the contaminant data in long format, both
the contaminants to be assessed and their auxiliary variables}

\item{info}{a harsat info object}

\item{determinands}{a character string given the identifiers of the
determinands that are to be assessed}
}
\value{
a data frame containing the contaminant data in wide format, with
the auxiliary variables pivoted to match the determinands they are linked to
}
\description{
Merge auxiliary variables with data
}
\details{
\code{info$determinand} identifies which auxiliary variables are linked to each
determinand; \code{info$auxiliary} also allows the user (currently limited) some
control over how the auxiliary variables are linked to the determinand data

Some variables can both be determinands to be assessed and auxiliary
variables (for example, CORG or DRYWT\%)
}
