% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsm_lmm.R
\name{negTwiceLogLik}
\alias{negTwiceLogLik}
\title{Utility function that returns double the negative log likelihood}
\usage{
negTwiceLogLik(data, mu, varComp)
}
\description{
Utility function that returns double the negative log likelihood
}
