% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{normalise_sediment_OSPAR}
\alias{normalise_sediment_OSPAR}
\title{Normalises sediment concentrations, OSPAR version}
\usage{
normalise_sediment_OSPAR(data, station_dictionary, info, control)
}
\arguments{
\item{data}{the data object}

\item{station_dictionary}{the station dictionary}

\item{info}{the information object}

\item{control}{control values}
}
\description{
Normalises sediment concentrations, OSPAR version
}
