% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportional_odds_functions.R
\name{ordinal_theta_cl}
\alias{ordinal_theta_cl}
\title{Calculates confidence limits for imposex time series}
\usage{
ordinal_theta_cl(fit, nsim = 1000, annual_indices = NULL)
}
\arguments{
\item{fit}{The output from a call to \code{ordinal_theta_est} (sort of)}

\item{nsim}{The number of simulations on which each set of confidence limits
is based; default 1000}
}
\description{
Calculates confidence limits for imposex time series
}
