% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportional_odds_functions.R
\name{ordinal_theta_est}
\alias{ordinal_theta_est}
\title{Estimates cut-points in ordinal model (imposex assessments)}
\usage{
ordinal_theta_est(data, theta = NULL, ref_level = NULL, calc_vcov = FALSE)
}
\arguments{
\item{data}{individual ordinal data (e.g. for imposex assessments)}

\item{theta}{an (optional) set of intial parameter values for theta; defaults
to NULL}

\item{ref_level}{an (optional) reference level for parameter estimation;
defaults to a index with intermediate levels of imposex}

\item{calc_vcov}{logical specifying whether to calculate the covariance
matrix of the parameter estimates; defaults to FALSE}
}
\description{
Estimates cut-points in ordinal model (imposex assessments)
}
