% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment_functions.R
\name{parallel_objects}
\alias{parallel_objects}
\title{Identifies functions needed for parallel processing in \code{run_assessment}}
\usage{
parallel_objects(imposex = FALSE)
}
\arguments{
\item{imposex}{A logical which, if TRUE, picks out extra functions required
for an imposex assessment; defaults to FALSE}
}
\value{
A vector of character strings of function names
}
\description{
Identifies functions needed for parallel processing in \code{run_assessment}
}
