% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_functions.R
\name{plot_assessment}
\alias{plot_assessment}
\title{Graphical summaries of an assessment}
\usage{
plot_assessment(
  assessment_obj,
  subset = NULL,
  output_dir = ".",
  file_type = c("data", "index", "auxiliary"),
  file_format = c("png", "pdf"),
  auxiliary = "default",
  control = list()
)
}
\arguments{
\item{assessment_obj}{An assessment object resulting from a call to
run_assessment}

\item{subset}{An optional vector specifying which timeseries are to be
plotted. An expression will be evaluated in the timeSeries component of
assessment_obj; use \code{series} to identify individual timeseries.}

\item{output_dir}{The output directory for the assessment plots (possibly
supplied using \code{file.path}). The default is the working directory. The
output directory must already exist.}

\item{file_type}{A character vector specifying the types of assessment plot.
The default \code{c("data", "index", "auxiliary")} produces three plots for
each time series. See details.}

\item{file_format}{A character string specifying Whether the files should be
png (the default) or pdf.}

\item{auxiliary}{A character string specifying the auxiliary variables
plotted if \code{file_type = "auxiliary"}. See details}

\item{control}{A list that can be used to modify the appearance of the plots.
See details.}
}
\value{
A series of png or pdf files with graphical summaries of an
assessment.
}
\description{
Generates a series of assessment plots for each time series. The plots are
exported as either png or pdf files.
}
\details{
\subsection{Types of assessment plots}{
\itemize{
\item \code{file_type = "data"} shows the raw data with the fitted trend and
pointwise two-sided 90\% confidence bands
\item \code{file_type = "index"} shows annual indices that summarise the data for
each year with the fitted trend and pointwise two-sided 90\% confidence
bands
\item \code{file_type = "auxiliary"} shows the raw data and key auxiliary variables
see below)
}
}

\subsection{Auxiliary variables}{

The default (\code{auxiliary = "default"}) is to plot the following variables:
\itemize{
\item biota: determinand concentration, LNMEA (mean length), DRYWT\% (dry weight
content), LIPIDWT\% (lipi weight content)
\item sediment: non-normalised determinand concentration, normalised
determinand concentration, AL (aluminium concentration), CORG (organic
carbon content)
\item water: no plots are generated at present
}

For biota, the determinand concentration will always be plotted, but it is
possible to change the three auxiliary variables. For example, to plot
WTMEA (mean weight) instead of LIPIDWT\% you would set \verb{auxiliary =  c("LNMEA", "WTMEA", "DRYWT\%)}. For this to work, WTMEA must previously have
been specified as an auxiliary variable for the determinand in question
using the \code{biota_auxliary} column in the determinand reference table. At
present, there must always be three auxiliary variables for biota.

For sediment, the non-normalised determinand concentration and the
normalised determinand concentration will always be plotted, but it is
possible to change the two auxiliary variables. For example, for metals in
sediment, you might set \code{auxiliary = c("AL", "LI")} to plot aluminium
and lithium concentrations instead of aluminium and organic carbon
concentrations. Again, for this to work, LI must previously have been
specified as an auxiliary variable for the determinand in question using
the \code{sediment_auxliary} column in the determinand reference table.
At present, there must always be two auxiliary variables for sediment.

At present, plots for only a limited range of auxiliary variables are
supported. More flexibility in these plots, such as changing the number of
auxiliary variables, is desirable and will emerge in due course.
}

\subsection{Control}{

At present the only aspect of the plots that can be changed is the range of
years that are shown on the x axis, for example by setting
\verb{control = list(xlim = c(2010, 2025)}. When multiple plots are produced,
specifying \code{control$xlim} will affect every plot.
}
}
