% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_functions.R
\name{pretty_unit}
\alias{pretty_unit}
\title{Pretty representation of units for plots and reports}
\usage{
pretty_unit(unit, html = FALSE)
}
\arguments{
\item{html}{A logical with TRUE returning an html representation for use in
markdown and FALSE (default) returning an text expression for use in lattice
(grid) graphics}
}
\value{
A vector of character strings that can be used in markdown (\code{html = TRUE}) or
in grid graphics ('html = FALSE\verb{). In the latter, need to turn the strings  into expressions by }parse(text = result)`.
}
\description{
Utilty function for adding units to plots in e.g. \code{plot_assessment} or
html reports in e.g. \code{report_assessment}. Standard (recognised) units are
prettified, with non-standard units returned unchanged.
}
