% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{read_data}
\alias{read_data}
\title{Read HARSAT data}
\usage{
read_data(
  compartment = c("biota", "sediment", "water"),
  purpose = c("OSPAR", "HELCOM", "AMAP", "custom"),
  contaminants,
  stations,
  data_dir = ".",
  data_format = c("ICES", "external"),
  info_files = list(),
  info_dir = ".",
  extraction = NULL,
  max_year = NULL,
  oddity_dir = "oddities",
  control = list(),
  file_diagnostics = FALSE
)
}
\arguments{
\item{compartment}{A string: \code{"biota"}, \code{"sediment"} or \code{"water"}}

\item{purpose}{A string specifying whether to use the default set up for
\code{"OSPAR"}, \code{"HELCOM"}, or \code{"AMAP"} or to use a customised setup \code{"custom"}}

\item{contaminants}{A file reference for the contaminant data}

\item{stations}{A file reference for the station data}

\item{data_dir}{The directory where the data files can be found (sometimes
supplied using 'file.path'). Defaults to "."; i.e. the working directory.}

\item{data_format}{A string specifying whether the data were extracted from
the ICES webservice (\code{"ICES"} - the default) or are in the simplified
format designed for other data sources (\code{"external"}).}

\item{info_files}{A list of files specifying reference tables which override
the defaults. See examples.}

\item{info_dir}{The directory where the reference tables can be found
(sometimes supplied using 'file.path'). Defaults to "."; i.e. the working
directory}

\item{extraction}{A date saying when the extraction was made. Optional. This
should be provided according to ISO 8601; for example, 29 February 2024
should be supplied as "2024-02-29". If the contaminant data were extracted
from the ICES webservice and the download file name has not been changed,
the extraction data will be taken from the contaminant file name.}

\item{max_year}{An integer giving the last monitoring year that should be
included in the assessment. Data from monitoring years after \code{max_year}
will be deleted. If not specified \code{max_year} is taken to be the last
monitoring year in the contaminant data file.}

\item{oddity_dir}{The directory where the 'oddities' will be written
(sometimes supplied using 'file.path'). This directory (and subdirectories)
will be created if it does not already exist.}

\item{control}{A list of control parameters that override the default values
used to run the assessment. These include the reporting window; the way in
which data are matched to stations following an ICES extraction;
information about reporting regions, and so on. See Details.}

\item{file_diagnostics}{A logical which, when \code{TRUE}, prints out the MD5
digest for each input file. Defaults to \code{FALSE}.}
}
\value{
A list with the following components:
\itemize{
\item \code{call} The function call.
\item \code{info} A list containing the reference tables and the control parameters.
\item \code{data} A data frame containing the contaminant (and effects) data. For
\code{external} data, this is identical to the input data file apart from some
extra empty columns which have been added. For \code{ICES} data, some existing
columns have been renamed (otherwise they are untouched) and some
additional columns have been constructed. The key ones of these are:
\itemize{
\item \code{station_code} the code of the station in the station dictionary that
best matches the data
\item \code{station_name} the name of the station
\item \code{species} (biota) the species based on \code{worms_accepted_name} where
available and \code{speci_name} otherwise
\item \code{filtration} (water) whether the sample was \code{filtered} or \code{unfiltered}
based on \code{method_pretreatment}
\item \code{retain} a logical indicating whether each record would have been
retained under the previous ICES extraction protocol. For example, \code{retain}
will be \code{FALSE} if the vflag entry is \code{"S"} or suspect. Records for which
\code{retain == FALSE} are deleted later in \code{tidy_data}
}
\item \code{stations}
}
}
\description{
Reads in contaminant and effects data, the station dictionary and various
reference tables. For data from the ICES webservice, it matches data to
stations in the station dictionary.  It also allows the user to set control
parameters that dictate the assessment process.
}
\details{
\subsection{Control parameters}{

Many aspects of the assessment process can be controlled using parameters
which are stored in the \code{info} component of the harsat data object. The
default control values can be overwritten using the \code{control} argument.
\itemize{
\item \code{reporting_window} A scalar (default 6) which determines whether
timeseries are excluded because they have no 'recent' data. Formally,
timeseries are excluded if they have no data in the period
\code{max_year - reporting_window + 1} and \code{max_year}, so the default approach
is to exclude timeseries if they have no dat in the most recent six
monitoring years. The value of 6 is chosen to match with Marine Strategy
Framework Directive reporting periods.
\item \code{region}
\item \code{add_stations}
\item \code{bivalve_spawning_season}
\item \code{use_stage}
\item \code{relative_uncertainty}
\item \code{auxiliary} A list which allows flexibility in the treatment of auxiliary
variables. At present, there is just one component \code{by_matrix}, a
character vector that determines which auxiliary variables are matched to
the contaminant data by \code{sample} and \code{matrix} as opposed to just \code{sample}.
For sediment and water, the default is \code{all}; i.e. all variables are
matched by \code{sample} and \code{matrix}. This ensures, for example, that
sediment normalisers such as aluminium and organic carbon content are
matched to chemical measurements in the same grain fraction. For biota, the
default is \verb{c("DRYWT\%", "LIPIDWT\%)}, so these variables are matched by
\code{sample} and \code{matrix} and all other variables (e.g. LNMEA or \%FEMALEPOP)
are matched by \code{sample}. Thus, dry weight and lipid weight contents are
matched to chemical measurements in the same tissue. However, mean length
(which is usually the lenght of the whole organism) is matched to all
tissue types.
}
}

\subsection{External data}{

If \code{data_format = "external"}, a simplified data and station file can
be supplied. See \code{vignette("external-file-format")} for details.
}
}
