% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{read_ICES_file}
\alias{read_ICES_file}
\title{Reads an ICES extraction}
\usage{
read_ICES_file(infile, var_id)
}
\arguments{
\item{infile}{The input file name}

\item{var_id}{A named character vector giving the name and type (character,
integer, logical, numeric, etc.) of each column in the file; these do not
need to be in the same order as in the file.}
}
\value{
A data frame containing the input file
}
\description{
Utility function to read an ICES extraction (either data or stations) having
specified the required column names and types. Prints a warning if required
column names are missing.
}
