% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{read_info_engine}
\alias{read_info_engine}
\title{Read a reference table}
\usage{
read_info_engine(id, files, compartment, file_diagnostics = FALSE)
}
\arguments{
\item{id}{Character string identifying the reference table; currently must
be one of \code{"determinand"}, \code{"species"}, \code{"thresholds"}, \code{"matrix"},
\code{"imposex"},\verb{"pivot_values", }"method_extraction"`.}

\item{files}{Named list of file paths, with one of the names matching \code{id}}

\item{compartment}{Character string: must be one of \code{"biota"}, \code{"sediment"}
or \code{"water"}}

\item{file_diagnostics}{Logical. The default (\code{FALSE}) is to print out a
simple message confirming which file has been read. When \code{TRUE}, the MD5
digest and full path is also printed.}
}
\value{
A reference table, or \code{NULL} if it is not required or cannot be
found.
}
\description{
Utility function for reading reference tables:
}
\details{
\itemize{
\item checks whether the table is required
\item checks whether the specified file exists
\item calls the appropriate function to read the file
\item generates messages, warnings and errors as appropriate
}
}
