% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_functions.R
\name{report_assessment}
\alias{report_assessment}
\title{Reports the assessment of individual time series}
\usage{
report_assessment(
  assessment_obj,
  subset = NULL,
  output_dir = ".",
  output_file = NULL,
  max_report = 100L
)
}
\arguments{
\item{assessment_obj}{An assessment object resulting from a call to
run_assessment}

\item{subset}{An optional vector specifying which timeseries are to be
reported. An expression will be evaluated in the timeSeries component of
assessment_obj; use 'series' to identify individual timeseries.}

\item{output_dir}{The output directory for the assessment plots (possibly
supplied using 'file.path'). The default is the working directory. The
output directory must already exist.}

\item{output_file}{An alterntive file name to override the default. This is
currently only implemented for a single report. If not supplied, the .html
extension will be added.}

\item{max_report}{The maximum number of reports that will be generated.
Defaults to 100. Each report is about 1MB in size and takes a few seconds
to run, so this prevents a ridiculous number of reports being created.}
}
\value{
A series of html files with, for each time series, meta data,
plots of the data with the fitted assessment model, statistical summaries,
and a simple interpretation of the fitted model.
}
\description{
Generates a series of html reports with, for each time series, meta data,
plots of the data with the fitted assessment model, statistical summaries,
and a simple interpretation of the fitted model.
}
