% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{report_file_digest}
\alias{report_file_digest}
\title{Generates and logs a file digest}
\usage{
report_file_digest(infile)
}
\arguments{
\item{file}{the file name}
}
\description{
Generates and logs a file digest
}
