% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment_functions.R
\name{run_assessment}
\alias{run_assessment}
\title{Assess timeseries for trends and status}
\usage{
run_assessment(
  ctsm_ob,
  subset = NULL,
  AC = NULL,
  get_AC_fn = NULL,
  recent_trend = 20L,
  parallel = FALSE,
  extra_data = NULL,
  control = list(),
  ...
)
}
\arguments{
\item{ctsm_ob}{A HARSAT object resulting from a call to create_timeSeries}

\item{subset}{An optional vector specifying which timeseries are to be
assessed. Might be used if the assessment is to be done in chunks because
of size, or when refitting a timeseries model which has not converged. An
expression will be evaluated in the timeSeries component of ctsm_ob; use
'series' to identify individual timeseries.}

\item{AC}{A character vector identifying the thresholds to be used in status
assessments. These should be in the threshold reference table. Defaults to
NULL; i.e. no thresholds are used.}

\item{get_AC_fn}{An optional function that overrides get_AC_default. See
details (which need to be written).}

\item{recent_trend}{An integer giving the number of years which are used in
the assessment of recent trends. For example, a value of 20 (the default)
consider trends in the last twenty year.}

\item{parallel}{A logical which determines whether to use parallel
computation; default = FALSE.}

\item{extra_data}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} A named list used to
pass additional data to specific assessment routines. At present it is
only used for imposex assessments, where it passes two data frames called
\code{VDS_estimates} and \code{VDS_confidence_limits}. Defaults to NULL, This
argument will be generalised in the near future, so expect it to change.}

\item{control}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} A list of control
parameters that allow the user to modify the way the assessment is run.
These currently include parameters for post-hoc power calculations and
for the calculation of the 'recent_change' (in addition to the
\code{recent_trend} argument above). See details.}

\item{...}{Extra arguments which are passed to assessment_engine. See
details (which need to be written).}
}
\description{
Fits a model to each timeseries, test for any temporal trend and compare with
thresholds. Need to add a lot more in details.
}
\details{
\subsection{Control parameters}{

Some aspects of the model output are controlled using parameters which
can be modified using the \code{control} argument. The default parameter list
is described below. The \code{control} argument only needs to specify any
changes. For example, to change the target power from 90\% (default) to 80\%,
use

\code{control = list(power = list(target_power = 80))}

The default parameters are a list with the following components:

\code{power}

A list with the following components (all expressed as
percentages):
\itemize{
\item \code{target_power} default = 90
\item \code{target_trend} default = 5
\item \code{size} default = 5
}

These affect the post-how power calculations. Power is currently only
calculated for time series of log-normally distributed data, which is why
the trend is expressed as a percentage.\if{html}{\out{
  <br>
  <br>
}}


\code{recent_change}

A list with the following components:
\itemize{
\item \code{n_year_fit} default = 5L
\item \code{n_year_positive} default = 5L
}

A recent change will only be computed if there are at least \code{n_year_fit}
years of data in the recent period, of which at least \code{n_year_positive}
contain at least one non-censored measurement. This only affects
normally or log-normally distributed data.

The default values of 5L mirror the requirements for calculating the
change over the whole time series.

Note that in harsat versions <= 1.0.2, \code{n_year_positive} was hard-wired to
0L which occasionally led to pathological behaviour in the estimation of
the recent change. To replicate previous outputs as closely as possible,
set \code{n_year_positive} to 2L (the smallest value allowed to avoid any
pathologial behavour). The change is only likely to affect long time series
with infrequent sampling where most measurements in the recent period are
censored.
}
}
