% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment_functions.R
\name{run_control_default}
\alias{run_control_default}
\title{Default control parameters for \code{run_assessment}}
\usage{
run_control_default()
}
\value{
A list with the following components:
\itemize{
\item \code{power} A list with the following components (all expressed as
percentages):
\itemize{
\item \code{target_power} default = 90\%
\item \code{target_trend} default = 5\%
\item \code{size} default = 5\%
The power calculations are currently only applied to log-normally
distributed data, which is why the trend is expressed as a percentage.
}
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Default parameters that control the way
the assessment is run. Presently only includes parameters for post-hoc
power, but it is intended to move \code{recent_trend} here, along with the
arguments that control the calculation of numerical derivatives.
}
