% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment_functions.R
\name{run_control_modify}
\alias{run_control_modify}
\title{Modifies control parameters for \code{run_assessment}}
\usage{
run_control_modify(control_default, control = list())
}
\arguments{
\item{control}{List of replacement control parameters; defaults to an empty
list.}

\item{run_control_default}{List of default control parameters produced by a
call to \code{run_control_default}}
}
\value{
List of updated control parameters
}
\description{
Undates default control parameters with user specification and does basis
error checking.
}
