% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{safe_read_file}
\alias{safe_read_file}
\title{Reads an input file, given a particular encoding, and possibly additional hints}
\usage{
safe_read_file(
  file,
  header = TRUE,
  sep = ",",
  quote = "\\"",
  dec = ".",
  fill = TRUE,
  comment.char = "",
  strip.white = TRUE,
  ...
)
}
\arguments{
\item{file}{the file name}

\item{...}{any additional parameters to \code{read.table}}
}
\value{
a tibble
}
\description{
Reads an input file, given a particular encoding, and possibly additional hints
}
