% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_functions.R
\name{subset_assessment}
\alias{subset_assessment}
\title{Subsets an assessment object}
\usage{
subset_assessment(assessment_obj, subset)
}
\arguments{
\item{assessment_obj}{An assessment object resulting from a call to
run_assessment.}

\item{subset}{A vector specifying the timeseries to be retained. An
expression will be evaluated in the timeSeries component of assessment_obj;
use 'series' to identify individual timeseries.}
}
\description{
Selects specific time series and simplifies the data, stations and
assessment components to match
}
