% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment_functions.R
\name{update_assessment}
\alias{update_assessment}
\title{Update timeseries assessments}
\usage{
update_assessment(ctsm_ob, subset = NULL, parallel = FALSE, ...)
}
\arguments{
\item{ctsm_ob}{A HARSAT object resulting from a call to run_assessment}

\item{subset}{A vector specifying which timeseries assessements are to be
updated or fit for the first time. An expression will be evaluated in the
timeSeries component of ctsm_ob; use 'series' to identify individual
timeseries.}

\item{parallel}{A logical which determines whether to use parallel
computation; default = FALSE.}

\item{...}{Extra arguments which are passed to assessment_engine.  See
details (which need to be written).}
}
\description{
Refits models for particular timeseries, or fits new models when an
assessment is being done in chunks.
}
