% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{add_stations}
\alias{add_stations}
\title{Add stations to contaminant data from an ICES extraction}
\usage{
add_stations(data, stations, info)
}
\arguments{
\item{data}{A data frame with the contaminant data from an ICES extraction}

\item{stations}{A data frame with the ICES station dictionary}

\item{info}{A HARSAT information list which must contain the elements
\code{purpose}, \code{compartment}, and \code{add_stations}. The latter is a list of
control parameters supplied through \code{control_default} or \code{control_modify}
which control how the station matching is achieved. See details.}
}
\value{
A data frame containing the contaminant data augmented by variables
containing the station code and the station name
}
\description{
Adds the station name and station code to the contaminant data from an ICES
extraction. This is done by either matching the station names submitted with
the data to the station dictionary, or by matching the sample coordinates to
the station dictionary, or a combination of both.
}
\details{
\code{info$add_stations} is a list of control parameters that modify the
station matching process:
\itemize{
\item method: a string specifying whether the stations are matched by \code{"name"},
\code{"coordinates"}, or \code{"both"}. If \code{info$purpose} is \code{"custom"}, \code{method} is
restricted to either \code{"name"} (the default) or \code{"coordinates"}. If
\code{info$purpose} is \code{"OSPAR"}, \code{"HELCOM"} or \code{"AMAP"}, then method is set to
\code{"both"} by default and stations are matched by name or coordinates
according to rules specified by OSPAR, HELCOM or AMAP data providers.
Specifically, stations are matched by name for Denmark, France (biota and
water - all years; sediment 2009 onwards), Ireland, Norway, Portugal, Spain
(2005 onwards), Sweden, The Netherlands (2007 onwards), United Kingdom. All
other stations are matched by coordinates.
\item area: a vector of strings containing one or more of \code{"OSPAR"}, \code{"HELCOM"}
and \code{"AMAP"}; this restricts the stations to those in the corresponding
convention area(s); NULL matches to all stations in the station dictionary
\item datatype: a logical specifying whether the stations should be restricted
to those with an appropriate datatype. If \code{TRUE}, a contaminant measurement
in biota (for example) will only be matched to stations with
\code{station_datatype} containing the string \code{"CF"}. Similarly, a biological
effect measurement in biota will only be matched to stations with
\code{station_datatype} containing the string \code{"EF"}
\item temporal: a logical with \code{TRUE} indicating that stations should be
restricted to those with \code{station_purpm} containing the string \code{"T"}
\item governance_type: a string: \code{"none"}, \code{"data"}, \code{"stations"} or \code{"both"}.
\code{"none"} means data and station governance are both ignored. \code{"data"} means
that matching will be restricted by data governance but not station
governance; for example if \code{governance_id == c("OSPAR", "AMAP")}, then data
will only be matched to a station if one of \code{is_ospar_monitoring} and
\code{is_amap_monitoring} is \code{TRUE}, with all stations considered regardless of
station governance. \code{"stations"} mean that matching will be restricted by
station governance but not by data governance; for example if
\code{governance_id == c("OSPAR", "AMAP")}, then the stations will be restricted
to those where \code{station_programgovernance} contains either \code{"OSPAR"} or
\code{"AMAP"}, with all data considered regardless of data governance. \code{both}
uses both data and station governance. If \code{governance_id} contains a single
value, then the matching is strict. However, if \code{governance_id} contains
multiple values, then the matching is more complicated. For example, if
\code{governance_id == c("OSPAR", "AMAP")}, then measurements with
\code{is_ospar_monitoring == TRUE} and \code{"is_amap_monitoring == FALSE"} are
matched to stations where \code{station_programgovernance} contains \verb{"OSPAR"; measurements with }is_ospar_monitoring == FALSE\code{and}is_amap_monitoring ==
TRUE\verb{are matched with stations where}station_programgovernance\code{contains}"AMAP"\verb{; but measurements where }is_ospar_monitoring == TRUE\code{and}is_amap_monitoring == TRUE\verb{are matched to stations where}station_programgovernance\verb{contains either}"OSPAR"\code{or}"AMAP"`.
\item governance_id: a vector of strings containing one or more of \code{"OSPAR"},
\code{"HELCOM"} and \code{"AMAP"}.
\item grouping: a logical with \code{TRUE} indicating that stations will be grouped
into meta-stations as specified by \code{station_asmtmimeparent} in the station
dictionary. Defaults to \code{FALSE} apart from when \code{info$purpose == "OSPAR"}.
\item check_coordinates: a logical with \code{TRUE} indicating that, when
stations are matched by name, the sample coordinates must also be within
the station geometry. No implemented yet, so defaults ot \code{FALSE}.
}
}
