% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{create_timeseries}
\alias{create_timeseries}
\title{Create a time series}
\usage{
create_timeseries(
  ctsm.obj,
  determinands = get_determinands(ctsm.obj$info),
  determinands.control = NULL,
  oddity_path = "oddities",
  return_early = FALSE,
  print_code_warnings = FALSE,
  get_basis = get_basis_default,
  normalise = FALSE,
  normalise.control = list()
)
}
\description{
Cleans the data and turns it into time series structures ready for assessment
}
