% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportional_odds_functions.R
\name{ctsm.VDS.loglik.calc}
\alias{ctsm.VDS.loglik.calc}
\title{ctsm.VDS.loglik.calc}
\usage{
ctsm.VDS.loglik.calc(
  theta,
  data,
  index.theta,
  minus.twice = FALSE,
  cumulate = FALSE
)
}
\arguments{
\item{cumulate}{a boolean, whether to use cumulative probabilities}
}
\description{
ctsm.VDS.loglik.calc
}
