% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment_functions.R
\name{ctsm_assessment}
\alias{ctsm_assessment}
\title{Run an assessment}
\usage{
ctsm_assessment(
  ctsm_ob,
  subset = NULL,
  AC = NULL,
  get_AC_fn = NULL,
  recent_trend = 20,
  parallel = FALSE,
  ...
)
}
\description{
Run an assessment
}
