% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{ctsm_read_data}
\alias{ctsm_read_data}
\title{Read harsat data}
\usage{
ctsm_read_data(
  compartment = c("sediment", "biota", "water"),
  purpose = c("OSPAR", "HELCOM", "AMAP", "other"),
  contaminants,
  stations,
  QA,
  data_path = ".",
  info_files = list(),
  info_path = ".",
  extraction = NULL,
  max_year = NULL,
  oddity_path = "oddities",
  control = list(),
  data_format = c("ICES_old", "ICES_new", "external")
)
}
\arguments{
\item{compartment}{a string, one of "sediment", "biota", and "water"}

\item{purpose}{a string, one of "OSPAR", "HELCOM", "AMAP" and "other"}

\item{contaminants}{a file reference for contaminant data}

\item{stations}{a file reference for station data}

\item{QA}{a file reference for QA data}

\item{data_path}{a file reference for data to read}
}
\description{
Read harsat data
}
