% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{ctsm_TBT_convert}
\alias{ctsm_TBT_convert}
\title{Convert tin concentrations}
\usage{
ctsm_TBT_convert(
  data,
  subset,
  action,
  from = c("tin", "cation"),
  convert_var = c("value", "limit_detection", "limit_quantification", "uncertainty")
)
}
\arguments{
\item{data}{the data}

\item{subset}{optional, a subset expression}

\item{action}{one of \code{"relabel"}, \code{"convert"}, \code{"change_unit"} --
note that \code{change_unit} moves units from tin units to conventional units}

\item{from}{either \code{"tin"} or \code{"cation"}}

\item{convert_var}{one of \code{"value"}, \code{"limit_detection"},
\code{"limit_quantification"}, \code{"uncertainty"}}
}
\description{
Convert tin concentrations to cation concentrations.
}
