% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_functions.R
\name{get_AC_biota_OSPAR}
\alias{get_AC_biota_OSPAR}
\title{Gets OSPAR threshold values for biota}
\usage{
get_AC_biota_OSPAR(data, AC, rt, export_all = FALSE)
}
\arguments{
\item{data}{}

\item{AC}{}

\item{rt}{}

\item{export_all}{}
}
\description{
Extends default extractor function get_AC$biota for the few cases which can
not be handled in a straightforward manner. This mostly relates to thresholds
which are only applied when the typical species / lipid contend is 'high'
(currently defined as >= 3\%)
}
