% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{get_timeseries}
\alias{get_timeseries}
\title{Extracts timeSeries}
\usage{
get_timeseries(harsat_obj, add = TRUE)
}
\arguments{
\item{harsat_obj}{A \code{harsat} object following a call to \code{\link{create_timeseries}}`.}

\item{add}{logical (default \code{TRUE}), if \code{TRUE}, adds extra information
about each station; if \code{FALSE}, simply returns the existing timeseries,}
}
\value{
A data.frame containing the timeSeries component with (optionally)
extra information about each station.
}
\description{
Gets the timeSeries component of a \code{harsat} object, optionally having added
extra information about each station
}
