% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harsat-package.R
\docType{package}
\name{harsat-package}
\alias{harsat}
\alias{harsat-package}
\title{harsat: Harmonized Regional Seas Assessment Tool}
\description{
Assessment of concentrations of hazardous substances and their biological effects in the marine environment. The code supports periodic international assessments by OSPAR, HELCOM and AMAP and also assessments by individual users. Includes tools for pre-processing data, statistical trend analysis and comparison with threshold values, and post-processing for archiving and reporting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/osparcomm/HARSAT}
  \item Report bugs at \url{https://github.com/osparcomm/HARSAT/issues}
}

}
\author{
\strong{Maintainer}: AmbieSense Ltd \email{info@ambiesense.com} [copyright holder]

Authors:
\itemize{
  \item Rob Fryer \email{rob.fryer@gov.scot}
  \item Leszek Kaliciak \email{leszek@ambiesense.com}
  \item Scottish Government \email{marinescotland@gov.scot} [funder]
}

Other contributors:
\itemize{
  \item Helsinki Commission (HELCOM) \email{secretariat@helcom.fi} [copyright holder, funder, contributor]
  \item Arctic Monitoring and Assessment Programme (AMAP) \email{amap@amap.no} [copyright holder, funder, contributor]
  \item OSPAR Commission (OSPAR) \email{secretariat@ospar.org} [copyright holder, funder, contributor]
  \item International Council for the Exploration of the Sea (ICES) \email{info@ices.dk} [copyright holder, contributor]
}

}
\keyword{internal}
