% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_functions.R
\docType{data}
\name{info_TEF}
\alias{info_TEF}
\title{TEF values for selected groups of compounds}
\format{
An object of class \code{list} of length 2.
}
\usage{
info_TEF
}
\description{
A list of named vectors which currently provide the TEFs for the WHO TEQ for
dioxins, furans and dioxin-like (planar) polychlorinated biphenyls.
DFP_environmental and DFP_human_health provide the TEFs appropriate for
testing against the environmental and human health standards respectively.
}
\details{
Adding further TEFs will require more development to the relevant
determinand_link function
}
\keyword{datasets}
