% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{locate_information_file}
\alias{locate_information_file}
\title{Find the path for an information file}
\usage{
locate_information_file(name, path)
}
\arguments{
\item{name}{A string: the name of the file, e.g., \code{thresholds_biota.csv}}

\item{path}{A vector of strings, directories to search. The information directory
for the package is automatically searched if we haven't found the
file anywhere else}
}
\value{
A string, the absolute path for the file, or \code{NULL} if the file
cannot be found anywhere.
}
\description{
Locates a requested information file, searching the
information file path. If the requested file cannot be found and
\code{required} is not false, stops with an error.
}
