% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_functions.R
\name{plot_assessment}
\alias{plot_assessment}
\title{Graphical summaries of an assessment}
\usage{
plot_assessment(
  assessment_obj,
  subset = NULL,
  output_dir = ".",
  file_type = c("data", "index"),
  file_format = c("png", "pdf")
)
}
\arguments{
\item{assessment_obj}{An assessment object resulting from a call to
run_assessment}

\item{subset}{An optional vector specifying which timeseries are to be
plotted. An expression will be evaluated in the timeSeries component of
assessment_obj; use 'series' to identify individual timeseries.}

\item{output_dir}{The output directory for the assessment plots (possibly
supplied using 'file.path'). The default is the working directory. The
output directory must already exist.}

\item{file_type}{Specifies whether the plots show the raw data ('file_type =
"data"'), annual indices summarising the data for each year ('file_type =
"index"'), or (the default) whether two files should be produced for each
time series, one with the raw data and one with the annual indices.}

\item{file_format}{Whether the files should be png (the default) or pdf.}
}
\value{
A series of png or pdf files with graphical summaries of an
assessment. The plots show the fitted trends with pointwise two-sided 90\%
confidence limits and either the raw data, or indices summarising the data
for each year.
}
\description{
Generates a series of assessment plots with the raw data, or the annual
indices, or both. The plots are exported as either png or pdf files.
}
