% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_functions.R
\name{plot_assessment}
\alias{plot_assessment}
\title{Graphical summaries of an assessment}
\usage{
plot_assessment(
  assessment_obj,
  subset = NULL,
  output_dir = ".",
  file_type = c("data", "index", "auxiliary"),
  file_format = c("png", "pdf"),
  auxiliary = "default"
)
}
\arguments{
\item{assessment_obj}{An assessment object resulting from a call to
run_assessment}

\item{subset}{An optional vector specifying which timeseries are to be
plotted. An expression will be evaluated in the timeSeries component of
assessment_obj; use \code{series} to identify individual timeseries.}

\item{output_dir}{The output directory for the assessment plots (possibly
supplied using \code{file.path}). The default is the working directory. The
output directory must already exist.}

\item{file_type}{A character vector specifying the types of assessment plot.
The default \code{c("data", "index", "auxiliary")} produces three plots for
each time series. See details}

\item{file_format}{A character string specifying Whether the files should be
png (the default) or pdf.}

\item{auxiliary}{A character string specifying the auxiliary variables
plotted if \code{file_type = "auxiliary"}. See details}
}
\value{
A series of png or pdf files with graphical summaries of an
assessment.
}
\description{
Generates a series of assessment plots for each time series. The plots are
exported as either png or pdf files.
}
\details{
\subsection{Types of assessment plots}{
\itemize{
\item \code{file_type = "data"} shows the raw data with the fitted trend and
pointwise two-sided 90\% confidence bands
\item \code{file_type = "index"} shows annual indices that summarise the data for
each year with the fitted trend and pointwise two-sided 90\% confidence
bands
\item \code{file_type = "auxiliary"} shows the raw data and key auxiliary variables
see below)
}
}

\subsection{Auxiliary variables}{

The default (\verb{auxiliary = "default}) is to plot the following variables:
\itemize{
\item biota: concentration, LNMEA (mean length), DRYWT\% (dry weight content),
LIPIDWT\% (lipi weight content)
\item sediment: non-normalised concentration, normalised concentration, AL
(aluminiun concentration), CORG (organic carbon content)
\item water: no plots are generated at present
}

At present, it is only possible to change the last three variable for biota
and the last two variables for sediment. For example, for metals in
sediment, you might specify \code{auxiliary = c("AL", "LI")} to plot aluminium
and lithium concentrations instead of aluminium and organic carbon.

The auxiliary variables must be specified as such in the determinand
reference table in e.g. the \code{biota_auxiliary} or \code{sediment_auxiliary}
column. At present, plots for only a limited range of auxiliary variables
are supported.
}
}
