% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{read_contaminants}
\alias{read_contaminants}
\title{Reads contaminant data}
\usage{
read_contaminants(file, data_dir = ".", info)
}
\arguments{
\item{file}{A file reference for the contaminant data.}

\item{data_dir}{A path to the directory holding the contaminant data.
Defaults to the working directory.}

\item{info}{A list containing at least the following two elements:
\itemize{
\item compartment: \code{"biota"}, \code{"sediment"} or \code{"water"}
\item data_format: \code{"ICES"} or \code{"external"}
}}
}
\value{
A data frame containing the contaminant data.
}
\description{
A quick way of reading the contaminant data without having to do any
station matching (if using \verb{data_format == "ICES})
}
