% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{read_data}
\alias{read_data}
\title{Read HARSAT data}
\usage{
read_data(
  compartment = c("biota", "sediment", "water"),
  purpose = c("OSPAR", "HELCOM", "AMAP", "custom"),
  contaminants,
  stations,
  QA,
  data_dir = ".",
  data_format = c("ICES", "external", "ICES_old", "ICES_new"),
  info_files = list(),
  info_dir = ".",
  extraction = NULL,
  max_year = NULL,
  oddity_dir = "oddities",
  control = list()
)
}
\arguments{
\item{compartment}{A string: \code{"biota"}, \code{"sediment"} or \code{"water"}}

\item{purpose}{A string specifying whether to use the default set up for
\code{"OSPAR"}, \code{"HELCOM"}, or \code{"AMAP"} or to use a customised setup \code{"custom"}}

\item{contaminants}{A file reference for the contaminant data}

\item{stations}{A file reference for the station data}

\item{QA}{A file reference for the QA data}

\item{data_dir}{The directory where the data files can be found (sometimes
supplied using 'file.path'). Defaults to "."; i.e. the working directory.}

\item{data_format}{A string specifying whether the data were extracted from
the ICES webservice ("ICES" - the default) or are in the simplified format
designed for other data sources ("external"). The values "ICES_old" and
"ICES_new" are deprecated.}

\item{info_files}{A list of files specifying reference tables which override
the defaults. See examples.}

\item{info_dir}{The directory where the reference tables can be found
(sometimes supplied using 'file.path'). Defaults to "."; i.e. the working
directory}

\item{extraction}{A date saying when the extraction was made. Optional. This
should be provided according to ISO 8601; for example, 29 February 2024
should be supplied as "2024-02-29". If the contaminant data were extracted
from the ICES webservice and the download file name has not been changed,
the extraction data will be taken from the contaminant file name.}

\item{max_year}{An integer giving the last monitoring year that should be
included in the assessment. Data from monitoring years after \code{max_year}
will be deleted. If not specified \code{max_year} is taken to be the last
monitoring year in the contaminant data file.}

\item{oddity_dir}{The directory where the 'oddities' will be written
(sometimes supplied using 'file.path'). This directory (and subdirectories)
will be created if it does not already exist.}

\item{control}{A list of control parameters that override the default values
used to run the assessment. These include the reporting window; the way in
which data are matched to stations following an ICES extraction;
information about reporting regions, and so on. See Details.}
}
\value{
A list with the following components:
\itemize{
\item \code{call}
\item \code{info}
\item \code{data}
\item \code{stations}
}
}
\description{
Reads in contaminant data, station dictionary and reference table. Also
allows the user to set various control parameters.
}
