% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{read_data}
\alias{read_data}
\title{Read HARSAT data}
\usage{
read_data(
  compartment = c("biota", "sediment", "water"),
  purpose = c("OSPAR", "HELCOM", "AMAP", "custom"),
  contaminants,
  stations,
  data_dir = ".",
  data_format = c("ICES", "external"),
  info_files = list(),
  info_dir = ".",
  extraction = NULL,
  max_year = NULL,
  oddity_dir = "oddities",
  control = list()
)
}
\arguments{
\item{compartment}{A string: \code{"biota"}, \code{"sediment"} or \code{"water"}}

\item{purpose}{A string specifying whether to use the default set up for
\code{"OSPAR"}, \code{"HELCOM"}, or \code{"AMAP"} or to use a customised setup \code{"custom"}}

\item{contaminants}{A file reference for the contaminant data}

\item{stations}{A file reference for the station data}

\item{data_dir}{The directory where the data files can be found (sometimes
supplied using 'file.path'). Defaults to "."; i.e. the working directory.}

\item{data_format}{A string specifying whether the data were extracted from
the ICES webservice (\code{"ICES"} - the default) or are in the simplified
format designed for other data sources (\code{"external"}).}

\item{info_files}{A list of files specifying reference tables which override
the defaults. See examples.}

\item{info_dir}{The directory where the reference tables can be found
(sometimes supplied using 'file.path'). Defaults to "."; i.e. the working
directory}

\item{extraction}{A date saying when the extraction was made. Optional. This
should be provided according to ISO 8601; for example, 29 February 2024
should be supplied as "2024-02-29". If the contaminant data were extracted
from the ICES webservice and the download file name has not been changed,
the extraction data will be taken from the contaminant file name.}

\item{max_year}{An integer giving the last monitoring year that should be
included in the assessment. Data from monitoring years after \code{max_year}
will be deleted. If not specified \code{max_year} is taken to be the last
monitoring year in the contaminant data file.}

\item{oddity_dir}{The directory where the 'oddities' will be written
(sometimes supplied using 'file.path'). This directory (and subdirectories)
will be created if it does not already exist.}

\item{control}{A list of control parameters that override the default values
used to run the assessment. These include the reporting window; the way in
which data are matched to stations following an ICES extraction;
information about reporting regions, and so on. See Details.}
}
\value{
A list with the following components:
\itemize{
\item \code{call} The function call.
\item \code{info} A list containing the reference tables and the control parameters.
\item \code{data} A data frame containing the contaminant (and effects) data. For
\code{external} data, this is identical to the input data file apart from some
extra empty columns which have been added. For \code{ICES} data, some existing
columns have been renamed (otherwise they are untouched) and some
additional columns have been constructed. The key ones of these are:
\itemize{
\item \code{station_code} the code of the station in the station dictionary that
best matches the data
\item \code{station_name} the name of the station
\item \code{species} (biota) the species based on \code{worms_accepted_name} where
available and \code{speci_name} otherwise
\item \code{filtration} (water) whether the sample was \code{filtered} or \code{unfiltered}
based on \code{method_pretreatment}
\item \code{retain} a logical indicating whether each record would have been
retained under the previous ICES extraction protocol. For example, \code{retain}
will be \code{FALSE} if the vflag entry is \code{"S"} or suspect. Records for which
\code{retain == FALSE} are deleted later in \code{tidy_data}
}
\item \code{stations}
}
}
\description{
Reads in contaminant and effects data, the station dictionary and various
reference tables. For data from the ICES webservice, it matches data to
stations in the station dictionary.  It also allows the user to set control
parameters that dictate the assessment process.
}
\details{
\subsection{Control parameters}{

Many aspects of the assessment process can be controlled through the
parameters stored in \code{info$control}. This is a list populated with default
values which can then be overwritten, if required, using the \code{control}
argument.
}

\subsection{External data}{

If \code{data_format = "external"}, a simplified data and station file can
be supplied. These should be .csv files, preferably saved with a UTF-8
encoding. All missing value should be supplied as empty cells, not as \code{NA}
or some other code.

The data file has one row for each measurement. The mandatory variables
are:
\itemize{
\item \code{country}: identifies the source of the data. For international
assessments, this is typically the country of origin, but for national
assessments it could be a local monitoring authority. It is read in as a
character string.
\item \code{station_code}: the code of the station where the data were collected.
This can be numeric, but is read in as a character string.
\item \code{station_name}: the name of the station where the data were collected.
This is usually more meaningful to a user than \code{station_code}.
\item \code{year}: an integer giving the monitoring year.
\item \code{sample}: an identifier that is used to match measurements from the same
individual (biota), sediment sample or water sample. This can be numeric,
but will be read in as a character string. Take care over this variable as
if it is incorrectly specified you could lose a lot of data.
\item \code{determinand}: the code identifying the thing measured. This should
match a record in the determinand reference table.
\item \code{matrix}: the code identifying the tissue (biota) or size fraction
(sediment) of the sample. Use ICES codes at present. For water it should
always be set to \code{"WT"}.
\item \code{unit}: the unit of the measurement. Use ICES codes at present.
\item \code{value}: the numeric value of the measurement.
}

The values of \code{country}, \code{station_code} and \code{station_name} (in each row)
must match an entry in the station file. No missing values are allowed in
the mandatory data variables.

The station file has one row for each station. The mandatory variables are:
\itemize{
\item \code{country}: see data file variables
\item \code{station_code}: see data file variables
\item \code{station_name}: see data file variables
\item \code{station_latitude}: the nominal latitude of the station (nominal because
samples are rarely taken at exactly the same location each year). This
should be provided as decimal degrees (or a numeric value based on some
other projection).
\item \code{station_longitude}: the nominal longitude of the station. This should
be provided as decimal degress (or a numeric value based on some other
projection).
}

The optional variables (including the semi-optional regional information)
will be documented at a later date.
}
}
