% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{read_stations}
\alias{read_stations}
\title{Reads station dictionary}
\usage{
read_stations(file, data_dir = ".", info)
}
\arguments{
\item{file}{A file reference for the station dictionary.}

\item{data_dir}{A path to the directory holding the station dictionary.
Defaults to the working directory.}

\item{info}{A list containing at least the following two elements:
\itemize{
\item compartment: \code{"biota"}, \code{"sediment"} or \code{"water"}
\item data_format: \code{"ICES"} or \code{"external"}
}}
}
\value{
A data frame containing the station dictionary.
}
\description{
Reads station dictionary
}
