% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{report_file_digest}
\alias{report_file_digest}
\title{Generates and logs a file digest}
\usage{
report_file_digest(infile)
}
\arguments{
\item{infile}{the input file name}
}
\description{
Writes a line on the console containing the file name and its
MD5 digest. This can be used to track changes in input files, for
reproducibility reasons. MD5 is good enough for this as we aren't
after cryptograohic levels of security, and its cheaper to compute.
}
