% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment_functions.R
\name{run_assessment}
\alias{run_assessment}
\title{Assess timeseries for trends and status}
\usage{
run_assessment(
  ctsm_ob,
  subset = NULL,
  AC = NULL,
  get_AC_fn = NULL,
  recent_trend = 20L,
  parallel = FALSE,
  extra_data = NULL,
  control = list(),
  ...
)
}
\arguments{
\item{ctsm_ob}{A HARSAT object resulting from a call to create_timeSeries}

\item{subset}{An optional vector specifying which timeseries are to be
assessed. Might be used if the assessment is to be done in chunks because
of size, or when refitting a timeseries model which has not converged. An
expression will be evaluated in the timeSeries component of ctsm_ob; use
'series' to identify individual timeseries.}

\item{AC}{A character vector identifying the thresholds to be used in status
assessments. These should be in the threshold reference table. Defaults to
NULL; i.e. no thresholds are used.}

\item{get_AC_fn}{An optional function that overrides get_AC_default. See
details (which need to be written).}

\item{recent_trend}{An integer giving the number of years which are used in
the assessment of recent trends. For example, a value of 20 (the default)
consider trends in the last twenty year.}

\item{parallel}{A logical which determines whether to use parallel
computation; default = FALSE.}

\item{extra_data}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} A named list used to
pass additional data to specific assessment routines. At present it is
only used for imposex assessments, where it passes two data frames called
\code{VDS_estimates} and \code{VDS_confidence_limits}. Defaults to NULL, This
argument will be generalised in the near future, so expect it to change.}

\item{control}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} A list of control
parameters that allow the user to modify the way the assessment is run.
At present, these only include parameters involved in post-hoc power
calculations, but it is intended to move other structures such as
\code{recent_trend} here. See details (which need to be written).}

\item{...}{Extra arguments which are passed to assessment_engine. See
details (which need to be written).}
}
\description{
Fits a model to each timeseries, test for any temporal trend and compare with
thresholds. Need to add a lot more in details.
}
