% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{safe_read_file}
\alias{safe_read_file}
\title{Reads an input file, given a particular encoding, and possibly additional hints}
\usage{
safe_read_file(
  file,
  header = TRUE,
  sep = ",",
  quote = "\\"",
  dec = ".",
  fill = TRUE,
  comment.char = "",
  strip.white = TRUE,
  ...
)
}
\arguments{
\item{file}{the file name}

\item{header}{logical (default \code{TRUE}), whether or not a header line is expected}

\item{sep}{character (defailt \verb{,}), field separator character}

\item{quote}{character (default \verb{"}), field quote character}

\item{dec}{character (default \code{.}), decimal character}

\item{fill}{logical (default \code{TRUE}), if rows have unequal lengtg, additional blank fields are added}

\item{comment.char}{character (default is empty), if not empty, a comment character}

\item{strip.white}{local (default \code{TRUE}), strips leading and trailing white space from fields}

\item{...}{any additional parameters to \code{\link[utils:read.table]{utils::read.table}}}
}
\value{
a data frame
}
\description{
Reads an input file, given a particular encoding, and possibly additional hints
}
