% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{tidy_data}
\alias{tidy_data}
\title{Tidy the data}
\usage{
tidy_data(ctsm_obj)
}
\arguments{
\item{ctsm_obj}{the time series data, typically returned from \code{\link{read_data}}.}
}
\description{
Reduces the size of the extraction by removing redundant variables.
Any ad-hoc changes will usually be made between \code{\link{read_data}} and \code{\link{tidy_data}}.
The output is in the correct format for \code{\link{create_timeseries}}.
}
