% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_functions.R
\name{write_summary_table}
\alias{write_summary_table}
\title{Write assessment summary to a csv file}
\usage{
write_summary_table(
  harsat_obj,
  output_file = NULL,
  output_dir = ".",
  export = TRUE,
  collapse_AC = NULL,
  extra_output = NULL,
  symbology = NULL,
  determinandGroups = NULL,
  append = FALSE
)
}
\arguments{
\item{harsat_obj}{A harsat object following a call to \code{run_assessment}.}

\item{output_file}{The name of the output csv file. If using NULL, the file
will be called \code{biota_summary.csv}, \code{sediment_summary.csv} or \code{water_summary.csv}
as appropriate. By default the file will be written to the working
directory. If a file name is provided, a path to the output file can also
be provided (e.g. using \code{file.path}). The `output_dir`` option can also be
used to specify the output file directory.}

\item{output_dir}{The output directory for \code{output_file}. The default is the
working directory. Any file path provided in \code{output_file}, will be
appended to \code{output_dir}. The resulting output directory must already
exist.}

\item{export}{Logical. \code{TRUE} (the default) writes the summary table to a csv
file. \code{FALSE} returns the summary table as an R object (and does not write to
a csv file).}

\item{collapse_AC}{A names list of valid assessment criteria that allows
assessment criteria of the same 'type' to be reported together. See
details.}

\item{extra_output}{A character vector specifying extra summary metrics
to be included in the output. Currently only recognises "power" to give the
seven power metrics computed for lognormally distributed data. Defaults to
\code{NULL}; i.e. no extra output.}

\item{symbology}{Experimental. Specifies the output symbology. Currently
assumes the thresholds are presented in increasing magnitude of
environmental risk.}

\item{determinandGroups}{optional, a list specifying \code{labels} and \code{levels}
to label the determinands. The life of this argument is limited.}

\item{append}{Logical. \code{FALSE} (the default) overwrites any existing summary
file. \code{TRUE} appends data to it, creating it if it does not yet exist.}
}
\value{
a summary object, when \code{export} is \code{FALSE}
}
\description{
Creates a data frame summarising the assessment of each time series and
writes it to a csv file. The summary includes:
\itemize{
\item meta-data such as the monitoring location and number of years of data for
each time series
\item the fitted values in the last monitoring year with associated upper
one-sided 95\% confidence limits
\item the trend assessments (p-values and trend estimates)
\item the status assessments (if there any thresholds)
\item (optionally) a symbology summarising the trend (shape) and status (colour)
of each time series. This is experimental.
}
}
